requirements.txt
nltk==3.8.1
matplotlib==3.5.2
pandas==2.0.2
pyarrow==12.0.1
numpy==1.24.3
ftfy==6.1.1
clip @ git+https://github.com/openai/CLIP.git
pillow==9.5.0
-i https://download.pytorch.org/whl/cu117
torch==2.0.1
torchvision==0.15.2
